Simplification of navMenu and breadcrumb
