+++ src/main/resources/META-INF/jahia-academy-template.tld
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+
+<taglib xmlns="http://java.sun.com/xml/ns/j2ee"
+        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+        xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-jsptaglibrary_2_0.xsd"
+        version="2.0">
+
+    <description>Academy functions</description>
+    <display-name>Academy functions</display-name>
+    <tlib-version>1.0.0-SNAPSHOT</tlib-version>
+    <short-name>af</short-name>
+    <uri>http://academy.jahia.org/af</uri>
+    <function>
+        <description>Finds the URL of the first sub-page under the given node</description>
+        <name>findFirstSubPageUrl</name>
+        <function-class>org.jahia.modules.academy.Functions</function-class>
+        <function-signature>java.lang.String findFirstSubPageUrl(org.jahia.services.content.JCRNodeWrapper)</function-signature>
+    </function>
+</taglib>
