--- src/main/resources/META-INF/jahia-academy-template.tld.disabled
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-
-<taglib xmlns="http://java.sun.com/xml/ns/j2ee"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-jsptaglibrary_2_0.xsd"
-        version="2.0">
-
-    <description> </description>
-    <display-name> </display-name>
-    <tlib-version>1.0.0-SNAPSHOT</tlib-version>
-    <short-name> </short-name>
-    <uri> </uri>
-    <function>
-        <description> </description>
-        <name> </name>
-        <function-class> </function-class>
-        <function-signature> </function-signature>
-        <example> </example>
-    </function>
-</taglib>
